[#43] rolling paper list 페이지, 내부 컴포넌트 분리 및 데이터 정렬
rolling paper list 페이지, 내부 컴포넌트 분리 및 데이터 정렬
